refactor: [#1211] move tracker tests to InMemoryTorrentRepository
