+++ src/core/torrent/repository/in_memory.rs
@@ -104,14 +104,80 @@ impl InMemoryTorrentRepository {
 
 #[cfg(test)]
 mod tests {
+    use std::net::{IpAddr, Ipv4Addr, SocketAddr};
     use std::sync::Arc;
 
+    use aquatic_udp_protocol::{AnnounceEvent, NumberOfBytes, PeerId};
+    use bittorrent_primitives::info_hash::fixture::gen_seeded_infohash;
+    use bittorrent_primitives::info_hash::InfoHash;
+    use torrust_tracker_configuration::TORRENT_PEERS_LIMIT;
+    use torrust_tracker_primitives::peer::Peer;
     use torrust_tracker_primitives::torrent_metrics::TorrentsMetrics;
+    use torrust_tracker_primitives::DurationSinceUnixEpoch;
 
     use crate::core::torrent::repository::in_memory::InMemoryTorrentRepository;
 
+    fn sample_info_hash() -> InfoHash {
+        "3b245504cf5f11bbdbe1201cea6a6bf45aee1bc0".parse::<InfoHash>().unwrap()
+    }
+
+    /// Sample peer whose state is not relevant for the tests
+    fn sample_peer() -> Peer {
+        complete_peer()
+    }
+
+    fn leecher() -> Peer {
+        incomplete_peer()
+    }
+
+    /// A peer that counts as `complete` is swarm metadata
+    /// IMPORTANT!: it only counts if the it has been announce at least once before
+    /// announcing the `AnnounceEvent::Completed` event.
+    fn complete_peer() -> Peer {
+        Peer {
+            peer_id: PeerId(*b"-qB00000000000000000"),
+            peer_addr: SocketAddr::new(IpAddr::V4(Ipv4Addr::new(126, 0, 0, 1)), 8080),
+            updated: DurationSinceUnixEpoch::new(1_669_397_478_934, 0),
+            uploaded: NumberOfBytes::new(0),
+            downloaded: NumberOfBytes::new(0),
+            left: NumberOfBytes::new(0), // No bytes left to download
+            event: AnnounceEvent::Completed,
+        }
+    }
+
+    /// A peer that counts as `incomplete` is swarm metadata
+    fn incomplete_peer() -> Peer {
+        Peer {
+            peer_id: PeerId(*b"-qB00000000000000000"),
+            peer_addr: SocketAddr::new(IpAddr::V4(Ipv4Addr::new(126, 0, 0, 1)), 8080),
+            updated: DurationSinceUnixEpoch::new(1_669_397_478_934, 0),
+            uploaded: NumberOfBytes::new(0),
+            downloaded: NumberOfBytes::new(0),
+            left: NumberOfBytes::new(1000), // Still bytes to download
+            event: AnnounceEvent::Started,
+        }
+    }
+
+    /// It generates a peer id from a number where the number is the last
+    /// part of the peer ID. For example, for `12` it returns
+    /// `-qB00000000000000012`.
+    fn numeric_peer_id(two_digits_value: i32) -> PeerId {
+        // Format idx as a string with leading zeros, ensuring it has exactly 2 digits
+        let idx_str = format!("{two_digits_value:02}");
+
+        // Create the base part of the peer ID.
+        let base = b"-qB00000000000000000";
+
+        // Concatenate the base with idx bytes, ensuring the total length is 20 bytes.
+        let mut peer_id_bytes = [0u8; 20];
+        peer_id_bytes[..base.len()].copy_from_slice(base);
+        peer_id_bytes[base.len() - idx_str.len()..].copy_from_slice(idx_str.as_bytes());
+
+        PeerId(peer_id_bytes)
+    }
+
     #[tokio::test]
-    async fn should_collect_torrent_metrics() {
+    async fn it_should_collect_torrent_metrics() {
         let in_memory_torrent_repository = Arc::new(InMemoryTorrentRepository::default());
 
         let torrents_metrics = in_memory_torrent_repository.get_torrents_metrics();
@@ -125,5 +191,133 @@ mod tests {
                 torrents: 0
             }
         );
+    }
+
+    #[tokio::test]
+    async fn it_should_return_74_peers_at_the_most_for_a_given_torrent() {
+        let in_memory_torrent_repository = Arc::new(InMemoryTorrentRepository::default());
+
+        let info_hash = sample_info_hash();
+
+        for idx in 1..=75 {
+            let peer = Peer {
+                peer_id: numeric_peer_id(idx),
+                peer_addr: SocketAddr::new(IpAddr::V4(Ipv4Addr::new(126, 0, 0, idx.try_into().unwrap())), 8080),
+                updated: DurationSinceUnixEpoch::new(1_669_397_478_934, 0),
+                uploaded: NumberOfBytes::new(0),
+                downloaded: NumberOfBytes::new(0),
+                left: NumberOfBytes::new(0), // No bytes left to download
+                event: AnnounceEvent::Completed,
+            };
+
+            let () = in_memory_torrent_repository.upsert_peer(&info_hash, &peer);
+        }
+
+        let peers = in_memory_torrent_repository.get_torrent_peers(&info_hash);
+
+        assert_eq!(peers.len(), 74);
+    }
+
+    #[tokio::test]
+    async fn it_should_return_the_peers_for_a_given_torrent_excluding_a_given_peer() {
+        let in_memory_torrent_repository = Arc::new(InMemoryTorrentRepository::default());
+
+        let info_hash = sample_info_hash();
+        let peer = sample_peer();
+
+        let () = in_memory_torrent_repository.upsert_peer(&info_hash, &peer);
+
+        let peers = in_memory_torrent_repository.get_peers_for(&info_hash, &peer, TORRENT_PEERS_LIMIT);
+
+        assert_eq!(peers, vec![]);
+    }
+
+    #[tokio::test]
+    async fn it_should_return_74_peers_at_the_most_for_a_given_torrent_when_it_filters_out_a_given_peer() {
+        let in_memory_torrent_repository = Arc::new(InMemoryTorrentRepository::default());
+
+        let info_hash = sample_info_hash();
+
+        let excluded_peer = sample_peer();
+
+        let () = in_memory_torrent_repository.upsert_peer(&info_hash, &excluded_peer);
+
+        // Add 74 peers
+        for idx in 2..=75 {
+            let peer = Peer {
+                peer_id: numeric_peer_id(idx),
+                peer_addr: SocketAddr::new(IpAddr::V4(Ipv4Addr::new(126, 0, 0, idx.try_into().unwrap())), 8080),
+                updated: DurationSinceUnixEpoch::new(1_669_397_478_934, 0),
+                uploaded: NumberOfBytes::new(0),
+                downloaded: NumberOfBytes::new(0),
+                left: NumberOfBytes::new(0), // No bytes left to download
+                event: AnnounceEvent::Completed,
+            };
+
+            let () = in_memory_torrent_repository.upsert_peer(&info_hash, &peer);
+        }
+
+        let peers = in_memory_torrent_repository.get_peers_for(&info_hash, &excluded_peer, TORRENT_PEERS_LIMIT);
+
+        assert_eq!(peers.len(), 74);
+    }
+
+    #[tokio::test]
+    async fn it_should_return_the_torrent_metrics() {
+        let in_memory_torrent_repository = Arc::new(InMemoryTorrentRepository::default());
+
+        let () = in_memory_torrent_repository.upsert_peer(&sample_info_hash(), &leecher());
+
+        let torrent_metrics = in_memory_torrent_repository.get_torrents_metrics();
+
+        assert_eq!(
+            torrent_metrics,
+            TorrentsMetrics {
+                complete: 0,
+                downloaded: 0,
+                incomplete: 1,
+                torrents: 1,
+            }
+        );
+    }
+
+    #[tokio::test]
+    async fn it_should_get_many_the_torrent_metrics() {
+        let in_memory_torrent_repository = Arc::new(InMemoryTorrentRepository::default());
+
+        let start_time = std::time::Instant::now();
+        for i in 0..1_000_000 {
+            let () = in_memory_torrent_repository.upsert_peer(&gen_seeded_infohash(&i), &leecher());
+        }
+        let result_a = start_time.elapsed();
+
+        let start_time = std::time::Instant::now();
+        let torrent_metrics = in_memory_torrent_repository.get_torrents_metrics();
+        let result_b = start_time.elapsed();
+
+        assert_eq!(
+            (torrent_metrics),
+            (TorrentsMetrics {
+                complete: 0,
+                downloaded: 0,
+                incomplete: 1_000_000,
+                torrents: 1_000_000,
+            }),
+            "{result_a:?} {result_b:?}"
+        );
+    }
+
+    #[tokio::test]
+    async fn it_should_return_the_peers_for_a_given_torrent() {
+        let in_memory_torrent_repository = Arc::new(InMemoryTorrentRepository::default());
+
+        let info_hash = sample_info_hash();
+        let peer = sample_peer();
+
+        let () = in_memory_torrent_repository.upsert_peer(&info_hash, &peer);
+
+        let peers = in_memory_torrent_repository.get_torrent_peers(&info_hash);
+
+        assert_eq!(peers, vec![Arc::new(peer)]);
     }
 }
